Change variable types in the minimum cover service interface

+++ src/main/java/com/github/matjanko/calculators/api/cover/MinimumCoverService.java
@@ -10,8 +10,8 @@ import org.springframework.stereotype.Service;
 @Service
 public interface MinimumCoverService {
 
-    double getCoverBySafeTransmissionOfBondForces(int barDiameter, int aggregateSize);
+    int getCoverBySafeTransmissionOfBondForces(int barDiameter, int aggregateSize);
 
-    double getCoverByProtectionAgainstCorrosion(ExposureClass exposureClass, StructuralClass structuralClass);
+    int getCoverByProtectionAgainstCorrosion(ExposureClass exposureClass, StructuralClass structuralClass);
 
 }
